+++ common-validation/src/main/java/cn/featherfly/common/validation/apt/ValidationProcessor.java
@@ -2,8 +2,7 @@ package cn.featherfly.common.validation.apt;
 
 import java.util.Set;
 
-import javax.annotation.processing.AbstractProcessor;
-import javax.annotation.processing.Messager;
+import javax.annotation.Nonnull;
 import javax.annotation.processing.ProcessingEnvironment;
 import javax.annotation.processing.RoundEnvironment;
 import javax.annotation.processing.SupportedAnnotationTypes;
@@ -12,29 +11,25 @@ import javax.annotation.processing.SupportedSourceVersion;
 import javax.lang.model.SourceVersion;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.TypeElement;
-import javax.tools.Diagnostic;
 
 import com.sun.source.tree.AnnotationTree;
 import com.sun.source.tree.Tree;
-import com.sun.tools.javac.api.JavacTrees;
 import com.sun.tools.javac.code.Flags;
-import com.sun.tools.javac.model.JavacElements;
-import com.sun.tools.javac.processing.JavacProcessingEnvironment;
 import com.sun.tools.javac.tree.JCTree;
 import com.sun.tools.javac.tree.JCTree.JCVariableDecl;
-import com.sun.tools.javac.tree.TreeMaker;
 import com.sun.tools.javac.tree.TreeTranslator;
-import com.sun.tools.javac.util.Context;
 import com.sun.tools.javac.util.List;
 import com.sun.tools.javac.util.Name;
-import com.sun.tools.javac.util.Names;
+
+import cn.featherfly.common.ast.JavacProcessor;
 
 @SupportedSourceVersion(SourceVersion.RELEASE_8)
-@SupportedAnnotationTypes({ "javax.validation.constraints.AssertFalse", "javax.validation.constraints.AssertTrue",
-        "javax.validation.constraints.Digits", "javax.validation.constraints.DecimalMax",
-        "javax.validation.constraints.DecimalMin", "javax.validation.constraints.Email",
-        "javax.validation.constraints.Future", "javax.validation.constraints.FutureOrPresent",
-        "javax.validation.constraints.Max", "javax.validation.constraints.Min", "javax.validation.constraints.Negative",
+@SupportedAnnotationTypes({ "javax.validation.Valid", "javax.validation.constraints.AssertFalse",
+        "javax.validation.constraints.AssertTrue", "javax.validation.constraints.Digits",
+        "javax.validation.constraints.DecimalMax", "javax.validation.constraints.DecimalMin",
+        "javax.validation.constraints.Email", "javax.validation.constraints.Future",
+        "javax.validation.constraints.FutureOrPresent", "javax.validation.constraints.Max",
+        "javax.validation.constraints.Min", "javax.validation.constraints.Negative",
         "javax.validation.constraints.NegativeOrZero", "javax.validation.constraints.NotBlank",
         "javax.validation.constraints.NotEmpty", "javax.validation.constraints.NotNull",
         "javax.validation.constraints.Null", "javax.validation.constraints.Past",
@@ -42,34 +37,21 @@ import com.sun.tools.javac.util.Names;
         "javax.validation.constraints.Positive", "javax.validation.constraints.PositiveOrZero",
         "javax.validation.constraints.Size" })
 @SupportedOptions("debug")
-public class ValidationProcessor extends AbstractProcessor {
-
-    private Messager messager;
-    private JavacTrees trees;
-    private TreeMaker treeMaker;
-    private Names names;
+public class ValidationProcessor extends JavacProcessor {
 
     @Override
-    public synchronized void init(ProcessingEnvironment processingEnv) {
-        super.init(processingEnv);
-        messager = processingEnv.getMessager();
-        trees = JavacTrees.instance(processingEnv);
-        Context context = ((JavacProcessingEnvironment) processingEnv).getContext();
-        treeMaker = TreeMaker.instance(context);
-        names = Names.instance(context);
+    public void doInit(@Nonnull ProcessingEnvironment processingEnv) {
     }
 
     @Override
     public boolean process(Set<? extends TypeElement> set, RoundEnvironment roundEnvironment) {
-        final JavacElements elementUtils = (JavacElements) processingEnv.getElementUtils();
-        Messager messager = processingEnv.getMessager();
-        messager.printMessage(Diagnostic.Kind.NOTE, "start generate validation");
-        messager.printMessage(Diagnostic.Kind.NOTE, "" + roundEnvironment.getRootElements().size());
+        debug("start generate validation");
+        notice("" + roundEnvironment.getRootElements().size());
 
         //        for (Element element : roundEnvironment.getElementsAnnotatedWith(NotNull.class)) {
         //            String name = element.getSimpleName().toString();
-        //            messager.printMessage(Diagnostic.Kind.NOTE, element.getKind() + " - " + name);
-        //            messager.printMessage(Diagnostic.Kind.NOTE,
+        //            notice( element.getKind() + " - " + name);
+        //            notice(
         //                    element.getEnclosingElement().getKind() + " - " + element.getEnclosingElement().getSimpleName());
         //        }
 
@@ -85,7 +67,7 @@ public class ValidationProcessor extends AbstractProcessor {
 
         for (Element element : roundEnvironment.getRootElements()) {
             String name = element.getSimpleName().toString();
-            messager.printMessage(Diagnostic.Kind.NOTE, element.getKind() + "-" + name);
+            notice(element.getKind() + "-" + name);
             JCTree jcTree = trees.getTree(element);
             jcTree.accept(new TreeTranslator() {
                 @Override
@@ -94,24 +76,20 @@ public class ValidationProcessor extends AbstractProcessor {
                     for (JCTree tree : jcClassDecl.defs) {
                         if (tree.getKind() == Tree.Kind.METHOD) {
                             JCTree.JCMethodDecl jcMethodDecl = (JCTree.JCMethodDecl) tree;
-                            messager.printMessage(Diagnostic.Kind.NOTE, "name: " + jcMethodDecl.name);
-                            //                            messager.printMessage(Diagnostic.Kind.NOTE, "parameters: " + jcMethodDecl.getParameters());
-                            //                            messager.printMessage(Diagnostic.Kind.NOTE,
+                            notice("name: " + jcMethodDecl.name);
+                            //                            notice( "parameters: " + jcMethodDecl.getParameters());
+                            //                            notice(
                             //                                    "typeParameters: " + jcMethodDecl.getTypeParameters());
                             for (JCVariableDecl jcVariableDecl : jcMethodDecl.getParameters()) {
-                                messager.printMessage(Diagnostic.Kind.NOTE,
-                                        "\tparam: " + jcVariableDecl.name.toString());
+                                notice("\tparam: " + jcVariableDecl.name.toString());
                                 for (AnnotationTree annotationTree : jcVariableDecl.mods.getAnnotations()) {
-                                    messager.printMessage(Diagnostic.Kind.NOTE,
-                                            "\t\tannotation: " + annotationTree.getAnnotationType());
-                                    messager.printMessage(Diagnostic.Kind.NOTE,
-                                            "\t\tannotation: " + annotationTree.getAnnotationType());
+                                    notice("\t\tannotation: " + annotationTree.getAnnotationType());
+                                    notice("\t\tannotation: " + annotationTree.getAnnotationType());
                                 }
                             }
                             if (jcMethodDecl.name.toString().equals("notNull")) {
-                                messager.printMessage(Diagnostic.Kind.NOTE,
-                                        jcMethodDecl.name + "\n" + jcMethodDecl.body.toString());
-                                messager.printMessage(Diagnostic.Kind.NOTE, "add System.out.println('Hello, world')");
+                                notice(jcMethodDecl.name + "\n" + jcMethodDecl.body.toString());
+                                notice("add System.out.println('Hello, world')");
                                 treeMaker.pos = jcMethodDecl.pos;
                                 jcMethodDecl.body = treeMaker
                                         .Block(0, List.of(
@@ -127,17 +105,16 @@ public class ValidationProcessor extends AbstractProcessor {
                                                                         List.<JCTree.JCExpression>of(
                                                                                 treeMaker.Literal("Hello, world!!!")))),
                                                 jcMethodDecl.body));
-                                messager.printMessage(Diagnostic.Kind.NOTE,
-                                        jcMethodDecl.name + "\n" + jcMethodDecl.body.toString());
+                                notice(jcMethodDecl.name + "\n" + jcMethodDecl.body.toString());
                             }
                         }
                     }
-                    //                    messager.printMessage(Diagnostic.Kind.NOTE, "****************************");
-                    //                    messager.printMessage(Diagnostic.Kind.NOTE, jcMethodDeclList.size() + "");
+                    //                    notice( "****************************");
+                    //                    notice( jcMethodDeclList.size() + "");
                     //                    jcMethodDeclList.forEach(jcMethodDecl -> {
-                    //                        messager.printMessage(Diagnostic.Kind.NOTE, jcMethodDecl.name + " has been processed");
+                    //                        notice( jcMethodDecl.name + " has been processed");
                     //                        JCMethodDecl setter = makeGetterMethodDecl(jcMethodDecl);
-                    //                        messager.printMessage(Diagnostic.Kind.NOTE, " setter name: " + setter.name);
+                    //                        notice( " setter name: " + setter.name);
                     //                        jcClassDecl.defs = jcClassDecl.defs.prepend(setter);
                     //                        jcClassDecl.defs = jcClassDecl.defs.prepend(makeSetterMethodDecl(jcMethodDecl));
                     //                    });
@@ -145,7 +122,7 @@ public class ValidationProcessor extends AbstractProcessor {
                 }
             });
         }
-        messager.printMessage(Diagnostic.Kind.NOTE, "end generate validation");
+        debug("end generate validation");
         return true;
     }
 
@@ -180,11 +157,4 @@ public class ValidationProcessor extends AbstractProcessor {
         String s = name.toString();
         return names.fromString("set" + s.substring(0, 1).toUpperCase() + s.substring(1, name.length()));
     }
-
-    private void log(String msg) {
-        if (processingEnv.getOptions().containsKey("debug")) {
-            processingEnv.getMessager().printMessage(Diagnostic.Kind.NOTE, msg);
-        }
-    }
-
 }