+++ common-db/change-log.md
@@ -1,7 +1,7 @@
 # 0.4.0 2021-12-29  
 1. MappingMode没去名称重构
 2. ObjectDbMixedMappingFactory实现统一使用数据库的列名进行映射，即jpa映射元数据的列名与数据库的列名大小写不一致（equals=false,equalsIgnoreCase=true）也会使用数据库的列名进行映射
-
+3. common-model相关的mapping类移动到common-db-mapping-model模块
 
 # 0.3.23 2021-12-19  
 1. 加入PlatformListJavaSqlTypeMapper,PlatformArrayJavaSqlTypeMapper,NumberListJavaSqlTypeMapper,NumberArrayJavaSqlTypeMapper,PlatformJavaSqlTypeStringMapper